requirements.txt
altair
jupyterlab>=1.0.0a3
pandas
pyarrow
pyowm
python-forecastio
shapely
matplotlib
descartes
pyowm
python-forecastio